backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
cryptography
bcrypt
pymysql
python-dotenv
pandas
scikit-learn
PyPDF2
python-docx
